+++ core/docs/changelog.md
@@ -1,4 +1,6 @@
 # TBD
+### Changes
+* Add a new user-friendly log message when setup or run timeout is exceeded during a test starting process
 
 # 1.15.0
 ### Changes